voice-agent/requirements.txt
livekit-agents
livekit-plugins-xai
supabase
python-dotenv
